+++ neural_compressor/experimental/common/model.py
@@ -17,7 +17,10 @@
 
 """common Model just collects the information to construct a Model."""
 
-from neural_compressor.model.model import get_model_fwk_name, MODELS, get_model_type
+import sys
+from neural_compressor.model.model import get_model_fwk_name, MODELS
+from neural_compressor.model.tensorflow_model import get_model_type
+from neural_compressor.utils import logger
 
 class Model(object):
     """A wrapper of the information needed to construct a Model."""
