+++ neural_compressor/experimental/quantization.py
@@ -28,7 +28,8 @@ from ..utils import logger
 from ..utils.utility import time_limit
 from ..utils.create_obj_from_config import create_dataloader
 from ..model import BaseModel
-from ..model.model import TensorflowQATModel, get_model_fwk_name
+from ..model.tensorflow_model import TensorflowQATModel
+from ..model.model import get_model_fwk_name
 from ..conf.config import QuantConf
 from ..conf.pythonic_config import Config
 from deprecated import deprecated
