+++ neural_compressor/metric/__init__.py
@@ -18,7 +18,7 @@
 
 """Intel Neural Compressor Metric."""
 
-from .metric import METRICS, BaseMetric, metric_registry
+from .metric import METRICS, Metric, BaseMetric, metric_registry
 from os.path import dirname, basename, isfile, join
 import glob
 
@@ -29,4 +29,4 @@ for f in modules:
         __import__(basename(f)[:-3], globals(), locals(), level=1)
 
 
-__all__ = ["METRICS", "BaseMetric", "metric_registry"]
+__all__ = ["METRICS", "Metric", "BaseMetric", "metric_registry"]
