+++ neural_compressor/metric/metric.py
@@ -31,6 +31,26 @@ tf = LazyImport('tensorflow')
 mx = LazyImport('mxnet')
 transformers = LazyImport('transformers')
 
+class Metric(object):
+    """A wrapper of the information needed to construct a Metric.
+
+    The metric class should take the outputs of the model as the metric's inputs,
+    neural_compressor built-in metric always take (predictions, labels) as inputs, it's
+    recommended to design metric_cls to take (predictions, labels) as inputs.
+    """
+    
+    def __init__(self, metric_cls, name='user_metric', **kwargs):
+        """Initialize a Metric with needed information.
+        
+        Args:
+            metric_cls (cls): Should be a sub_class of neural_compressor.metric.BaseMetric, 
+                which takes (predictions, labels) as inputs
+            name (str, optional): Name for metric. Defaults to 'user_metric'.
+        """
+        self.metric_cls = metric_cls
+        self.name = name
+        self.kwargs = kwargs
+
 @singleton
 class TensorflowMetrics(object):
     """Tensorflow metrics collection.
