+++ neural_compressor/model/__init__.py
@@ -15,7 +15,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .model import MODELS, BaseModel
+from .model import MODELS, Model
+from .base_model import BaseModel
 
-__all__ = ["MODELS", "BaseModel"]
+__all__ = ["MODELS", "Model", "BaseModel"]
+
